+++ mudicom/exceptions.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+"""
+    mudicom.exceptions
+    ~~~~~~~~~~~~~~~~~~
+
+    Module for package specific exceptions
+"""
+
+
+class InvalidDicom(IOError):
+    pass